Create bridge_controller.ino
monitors a file, gets commands to open/close at different speeds from it. Linino side writes to the same file. 

Having a file is better in case reboots/crashes of components happen. There are Mailox and similar methods but they are essentially the same with some improvements I am not considering right now.